refac: divide response, request
