+++ src/main/java/nextstep/controller/MemberController.java
@@ -2,8 +2,8 @@ package nextstep.controller;
 
 import auth.domain.UserDetails;
 import auth.support.template.LoginMember;
-import nextstep.controller.dto.MemberRequest;
-import nextstep.controller.dto.MemberResponse;
+import nextstep.controller.dto.request.MemberRequest;
+import nextstep.controller.dto.response.MemberResponse;
 import nextstep.service.MemberService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -27,6 +27,11 @@ public class MemberController {
 
     @GetMapping("/me")
     public ResponseEntity<?> me(@LoginMember UserDetails userDetails) {
-        return ResponseEntity.ok(new MemberResponse(userDetails));
+        return ResponseEntity.ok(
+                new MemberResponse(userDetails.getUsername(),
+                        userDetails.getName(),
+                        userDetails.getPhone(),
+                        userDetails.getRole()
+        ));
     }
 }
