+++ src/main/java/nextstep/controller/WaitingController.java
@@ -3,8 +3,8 @@ package nextstep.controller;
 import auth.domain.UserDetails;
 import auth.support.template.LoginMember;
 import lombok.RequiredArgsConstructor;
-import nextstep.controller.dto.WaitingRequest;
-import nextstep.controller.dto.WaitingResponse;
+import nextstep.controller.dto.request.WaitingRequest;
+import nextstep.controller.dto.response.WaitingResponse;
 import nextstep.service.WaitingService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
