--- src/main/java/nextstep/controller/dto/ReservationRequest.java
@@ -1,16 +1,0 @@
-package nextstep.controller.dto;
-
-public class ReservationRequest {
-    private long scheduleId;
-
-    public ReservationRequest() {
-    }
-
-    public ReservationRequest(long scheduleId) {
-        this.scheduleId = scheduleId;
-    }
-
-    public long getScheduleId() {
-        return scheduleId;
-    }
-}
