--- src/main/java/nextstep/controller/dto/ThemeRequest.java
@@ -1,35 +1,0 @@
-package nextstep.controller.dto;
-
-import nextstep.domain.Theme;
-
-public class ThemeRequest {
-    private String name;
-    private String desc;
-    private int price;
-
-    public ThemeRequest(String name, String desc, int price) {
-        this.name = name;
-        this.desc = desc;
-        this.price = price;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getDesc() {
-        return desc;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public Theme toEntity() {
-        return new Theme(
-                this.name,
-                this.desc,
-                this.price
-        );
-    }
-}
