--- src/main/java/nextstep/controller/dto/WaitingRequest.java
@@ -1,10 +1,0 @@
-package nextstep.controller.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class WaitingRequest {
-    private long scheduleId;
-}
