+++ src/main/java/nextstep/controller/dto/request/ReservationRequest.java
@@ -1,0 +1,13 @@
+package nextstep.controller.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+@AllArgsConstructor
+public class ReservationRequest {
+    private long scheduleId;
+}
