+++ src/main/java/nextstep/controller/dto/request/WaitingRequest.java
@@ -1,0 +1,10 @@
+package nextstep.controller.dto.request;
+
+import lombok.*;
+
+@Getter
+@RequiredArgsConstructor
+@AllArgsConstructor
+public class WaitingRequest {
+    private long scheduleId;
+}
