+++ src/main/java/nextstep/service/ThemeService.java
@@ -1,6 +1,6 @@
 package nextstep.service;
 
-import nextstep.controller.dto.ThemeRequest;
+import nextstep.controller.dto.request.ThemeRequest;
 import nextstep.domain.Theme;
 import nextstep.repository.ThemeDao;
 import org.springframework.stereotype.Service;
