+++ src/test/java/nextstep/theme/ThemeDaoTest.java
@@ -2,6 +2,7 @@ package nextstep.theme;
 
 import nextstep.domain.Theme;
 import nextstep.repository.ThemeDao;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
@@ -14,6 +15,7 @@ public class ThemeDaoTest {
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
+    @DisplayName("테마를 생성한다")
     @Test
     void save() {
         ThemeDao themeDao = new ThemeDao(jdbcTemplate);
